% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainModel.R
\name{MainModel}
\alias{MainModel}
\title{Train a model with main and interaction effects}
\usage{
MainModel(
  data_encoded,
  main_vars,
  interaction_vars,
  response,
  learning_rate = NULL,
  max_iter_main = 1000,
  max_iter_interaction = 1000,
  patience = 5
)
}
\arguments{
\item{data_encoded}{A data frame containing the dataset.}

\item{main_vars}{A vector of column names representing the main effect variables.}

\item{interaction_vars}{A vector of column names representing the interaction effect variables.}

\item{response}{A string specifying the name of the response variable.}

\item{learning_rate}{A numeric value controlling the learning rate.}

\item{max_iter_main}{An integer specifying the maximum iterations for main effects (default = 1000).}

\item{max_iter_interaction}{An integer specifying the maximum iterations for interaction effects (default = 1000).}

\item{patience}{An integer specifying the early stopping patience (default = 5).}
}
\value{
A data frame containing the estimated effects for main and interaction variables.
}
\description{
Train a model with main and interaction effects
}
\keyword{internal}
