% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{perform_permutation_test}
\alias{perform_permutation_test}
\title{Internal function for the library NpmAnova (Perform a Permutation Test to Compute P-Values)}
\usage{
perform_permutation_test(
  org_model,
  df_encoded,
  response,
  main_vars_Full,
  interaction_vars_Full,
  main_var_inp,
  interaction_var_inp,
  n = 100,
  learning_rate
)
}
\arguments{
\item{org_model}{A trained model obtained from \code{MainModel()}.}

\item{df_encoded}{A data frame containing the dataset with encoded variables.}

\item{response}{A string specifying the response variable name.}

\item{main_vars_Full}{A character vector of all main effect variables.}

\item{interaction_vars_Full}{A character vector of all interaction effect variables.}

\item{main_var_inp}{A character vector of main effect variables to be included in the reduced model.}

\item{interaction_var_inp}{A character vector of interaction effect variables to be included in the reduced model.}

\item{n}{An integer specifying the number of permutations (default = 100).}

\item{learning_rate}{A numeric vector for the learning rates.}
}
\value{
A numeric p-value representing the statistical significance of the removed variable.
}
\description{
Evaluates the significance of an effect by running a permutation test
and comparing the model performance with and without the variable.
}
\keyword{internal}
