% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predicted_values}
\alias{predicted_values}
\title{Internal function for the library NpmAnova}
\usage{
predicted_values(df_encoded, variables, final_model)
}
\arguments{
\item{df_encoded}{A data frame containing the dataset with encoded variables.}

\item{variables}{A character vector of variable names used in the model.}

\item{final_model}{A data frame containing the estimated effects for main and interaction variables.}
}
\value{
A numeric vector of predicted values.
}
\description{
Computes predictions for a given dataset using the model
containing main and interaction effects.
}
\keyword{internal}
