% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{select_best_variable}
\alias{select_best_variable}
\title{Internal function for the library NpmAnova}
\usage{
select_best_variable(data, variables, z)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{variables}{A character vector of variable names to evaluate.}

\item{z}{A numeric vector of adjusted residuals from the model.}
}
\value{
A string representing the variable that results in the smallest SSE.
}
\description{
Identifies the variable that minimizes the sum of squared errors (SSE)
when used as a binary split in the tree.
}
\keyword{internal}
